+++ web/src/components/organisms/Project/Asset/Asset/index.tsx
@@ -12,6 +12,7 @@ const Asset: React.FC = () => {
   const { assetId } = useParams();
   const {
     asset,
+    assetUrl,
     assetFileExt,
     isLoading,
     selectedPreviewType,
@@ -23,6 +24,7 @@ const Asset: React.FC = () => {
     isSaveDisabled,
     updateLoading,
     hasUpdateRight,
+    setAssetUrl,
     handleAssetDecompress,
     handleAssetItemSelect,
     handleAssetDownload,
@@ -53,6 +55,7 @@ const Asset: React.FC = () => {
         />
       }
       asset={asset}
+      assetUrl={assetUrl}
       assetFileExt={assetFileExt}
       selectedPreviewType={selectedPreviewType}
       isModalVisible={isModalVisible}
@@ -62,6 +65,7 @@ const Asset: React.FC = () => {
       isSaveDisabled={isSaveDisabled}
       updateLoading={updateLoading}
       hasUpdateRight={hasUpdateRight}
+      setAssetUrl={setAssetUrl}
       onAssetItemSelect={handleAssetItemSelect}
       onAssetDecompress={handleAssetDecompress}
       onAssetDownload={handleAssetDownload}
